Added screenshots and implemented buffer

+++ Lab-4/Libs/buffer_fifo.h
@@ -3,7 +3,7 @@
 #ifndef BUFFER_FIFO
 #define BUFFER_FIFO
 
-typedef enum {
+typedef enum { // BE == Buffer Error
 	BE_EMPTY = -4,
 	BE_FULL = -3,
 	BE_NULL_PTR = -2,
@@ -12,11 +12,11 @@ typedef enum {
 } bufError_T;
 
 typedef struct CircBuf{
-	uint8_t *head;
-	uint8_t *tail;
+	uint32_t *head;
+	uint32_t *tail;
 	volatile uint32_t num_items;
 	uint32_t length;
-	uint8_t *buffer;
+	uint32_t *buffer;
 } CircBuf_T;
 
 bufError_T initializeBuffer(CircBuf_T *buf, uint32_t length);
@@ -24,7 +24,7 @@ void clearBuffer(CircBuf_T *buf);
 void deleteBuffer(CircBuf_T *buf);
 int8_t bufferFull(CircBuf_T *buf);
 int8_t bufferEmpty(CircBuf_T *buf);
-bufError_T addItem(CircBuf_T *buf, uint8_t item);
-bufError_T removeItem(CircBuf_T *buf, uint8_t *val);
+bufError_T addItem(CircBuf_T *buf, uint32_t item);
+bufError_T removeItem(CircBuf_T *buf, uint32_t *val);
 
 #endif
